リファクタリングのための修正

+++ app/controllers/goods_controller.rb
@@ -9,8 +9,7 @@ class GoodsController < ApplicationController
 
   def destroy
     good = Good.find(params[:id])
-    redirect_to root_path and return unless current_user == good.user 
-    good.destroy
+    good.destroy if current_user == good.user
     redirect_to room_path(params[:room_id])
   end
